add util.ts for sharing common functions between frontend and backend
