+++ package.json
@@ -56,6 +56,7 @@
         "vue-toastification": "^2.0.0-rc.1"
     },
     "devDependencies": {
+        "@types/bootstrap": "^5.0.17",
         "@vitejs/plugin-legacy": "^1.5.0",
         "@vitejs/plugin-vue": "^1.3.0",
         "@vue/compiler-sfc": "^3.1.5",
@@ -66,6 +67,7 @@
         "stylelint": "^13.13.1",
         "stylelint-config-recommended": "^5.0.0",
         "stylelint-config-standard": "^22.0.0",
+        "typescript": "^4.3.5",
         "vite": "^2.4.4"
     }
 }
