+++ server/server.js
@@ -12,7 +12,7 @@ const { getSettings } = require("./util-server");
 const { Notification } = require("./notification")
 const gracefulShutdown = require("http-graceful-shutdown");
 const Database = require("./database");
-const { sleep } = require("./util");
+const { sleep } = require("../src/util");
 const args = require("args-parser")(process.argv);
 const prometheusAPIMetrics = require("prometheus-api-metrics");
 const { basicAuth } = require("./auth");
